Fixed roll and pitch functions

+++ Project/hardware/inc/gyroCompass.h
@@ -28,8 +28,8 @@ uint8_t gyroscope_get_acceleration(int16_t aAcceleration[3]);
 uint8_t gyroscope_get_gyro(float aGyro[3]);
 
 //Useful data functions
-uint8_t gyroscope_get_roll(int16_t *aRoll);
-uint8_t gyroscope_get_pitch(int16_t *aPitch);
+uint8_t gyroscope_get_roll(int *aRoll);
+uint8_t gyroscope_get_pitch(int *aPitch);
 uint8_t gyroCompass_get_heading(int *aHeading);
 
 #endif /* GYROCOMPASS_H */
